+++ src/Components/GenshinImpact/data/IGOOD.ts
