+++ src/Components/GenshinImpact/data/WeaponKey.ts
